requirements.txt
coverage==3.7.1
Django==1.6.5
djangorestframework==2.3.13
django-dynamic-fixture==1.7.0
django-nose==1.2
django-tastypie==0.11.0
flake8==2.1.0
inflection==0.2.0
mock==1.0.1
python-mimeparse==0.1.4
PyYAML==3.10
requests==2.3.0
South==0.8.4
wsgiref==0.1.2
